requirements.txt
pandas>=1.2.2
matplotlib>=3.3.3
opencv-python>=4.5.1.0
seaborn>=0.11.2
fastprogress>=1.0.0
torch-summary>=1.4.5
tqdm>=4.35.0
wandb>=0.10.25
scikit-image>=0.18.1
tensorboardX==2.2
resnest>=0.0.5
xgboost>=1.4.2
kornia>=0.5.3
pydicom>=2.1.2
tensorflow==2.4.1
sklearn
pycocotools>=2.0.2
timm>=0.4.9
efficientnet-pytorch==0.6.3
pretrainedmodels>=0.7.4
plotly>=5.1.0
fastai>=2.4.0
SimpleITK>=2.1.0
torchio>=0.18.54
optuna>=2.9.1
xgboost>=1.7.1
wandb>=0.12.0